Fix setting response status code
