+++ src/utils/helpers.ts
@@ -1,6 +1,3 @@
-import { OverlayData, OverlayTeam } from "./types";
-import { v } from "./validator";
-
 /** Generate random id of given length between 4 and 32, defaults to 16 */
 export const generateRandomId = (length = 16) => {
   // Ensure length is within bounds
@@ -16,26 +13,4 @@ export const generateRandomId = (length = 16) => {
   return Math.random()
     .toString(36)
     .substring(2, length + 2);
-};
-
-/**
- * @throws Error if data does not match schema
- */
-export const validateOverlayData = (data: OverlayData) => {
-  const teamSchema = v.object({
-    score: v.number().integer().min(0),
-    name: v.string().isNotEmpty(),
-    primaryColor: v.string().isNotEmpty(),
-    secondaryColor: v.string().isNotEmpty(),
-    logoUrl: v.string().isNotEmpty(),
-  });
-
-  const overlaySchema = v.object({
-    maxScore: v.number().integer().min(1),
-    blue: teamSchema,
-    red: teamSchema,
-    cameraControlsCover: v.boolean(),
-  });
-
-  overlaySchema.parse(data);
 };
